[server:18447591]  change functions for user, token to class
