+++ server/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "KIT_server",
-  "version": "0.3.0",
+  "name": "kit_server",
+  "version": "0.4.0",
   "description": "",
   "main": "dist/index.js",
   "private": true,
