+++ server/src/database/functions/index.ts
@@ -1,2 +1,2 @@
-export * from './user.fn';
-export * from './token.fn';
+export * from './user.class';
+export * from './token.class';
